requirements.txt
python-telegram-bot==20.6
pandas==2.1.4
matplotlib==3.8.2
seaborn==0.13.0
